requirements.txt
pandas>=1.3.0
numpy>=1.21.0
ta>=0.7.0
python-binance>=1.0.15
ccxt>=2.0.0
python-telegram-bot>=13.7
python-dotenv>=0.19.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
asyncio>=3.4.3
aiohttp>=3.7.4
